// src/types/messages.ts
export interface Message {
  id?: string;
  createdAt?: string;
  message?: string;
  receiverId?: string;
  senderId?: string;
  senderName?: string;
  productId?: string;
  productTitle?: string;
  productImage?: string;
  last_updated?: string;
  product_id?: string;
  buyer_id?: string;
  last_message?: string;
}